them sale

+++ src/garden/garden.service.ts
@@ -8,62 +8,94 @@ export class GardenService{
     constructor (private readonly prisma : PrismaService){}
 
     // create garden
-    async createGarden(dto: CreateGardenDto, userId: number){
-       // kiểm tra tên trùng lặp
-       const existingGarden = await this.prisma.garden.findFirst({
-        where:{name: dto.name,userId}
-       })
-       if (existingGarden) {
-        throw new BadRequestException('Garden name already exists for this user');
-      }
+    async createGarden(dto: CreateGardenDto, user: any){
+        let userId: any
 
-      try {
+        //nếu là admin và truyền vào userId thì cho phép
+        if(user.role === 'admin' && dto.userId){
+            const user = await this.prisma.user.findUnique({
+                where: { id: dto.userId },
+              });
+              if (!user) {
+                throw new NotFoundException('User không tồn tại');
+              }
+
+            userId = dto.userId
+        } 
+        else{
+            if(dto.userId && dto.userId !== user.id){
+                throw new ForbiddenException('You are not allowed to create a garden for another user');
+            }
+
+            userId = user.id
+        }
+
+       // kiểm tra tên trùng lặp
+        const existingGarden = await this.prisma.garden.findFirst({
+            where:{name: dto.name, userId}
+        })
+        if (existingGarden) {
+            throw new BadRequestException('Garden name already exists for this user');
+        }
+
         return await this.prisma.garden.create({
-          data: {
-            name: dto.name,
-            userId // Liên kết với user
-          },
-        });
-      } catch (error) {
-        throw new BadRequestException('Failed to create garden');
-      }
+            data:{
+                name: dto.name,
+                userId
+            }
+        })
     }
 
-    // get all garden 
-    async getAllGarden(){
-        return await this.prisma.garden.findMany()
+    //get garden by userId
+    async getGarden(user: any){
+        if(user.role === 'admin'){
+            return await this.prisma.garden.findMany()
+        } else{
+            return await this.prisma.garden.findMany({
+                where:{userId: user.id}
+            })
+        }
     }
 
     //get garden by id
-    async getGardenById(id: number, userIdToken: number){
+    async getGardenById(id: number, user: any){
         const garden = await this.prisma.garden.findUnique({
             where: {id}
         })
         if(!garden){
             throw new NotFoundException(`Not found garden: ${id}`)
-        }else if(garden.userId != userIdToken){
+        }
+        
+        if(user.role !== 'admin' && garden.userId != user.id){
             throw new ForbiddenException("You are not authorized to access this resource")
         }
         return garden
     }
 
-    //get garden by userId
-    async getGardenByUserId(userId: number){
-        const gardens = await this.prisma.garden.findMany({
-            where:{userId}
-        })
-        if(!gardens){
-            throw new NotFoundException(`Not found userId:${userId}`)
-        } 
-        return gardens
-    }
-
     //update garden 
-    async updateGardenById(id: number, dto: UpdateGardenDto){
+    async updateGardenById(id: number, dto: UpdateGardenDto, user: any ){
         const garden = await this.prisma.garden.findUnique({where:{id}})
         if(!garden){
             throw new NotFoundException(`Not found garden: ${id}`)
         }
+
+        // Chỉ cho admin hoặc chính chủ sở hữu garden cập nhật
+        if (user.role !== 'admin' && garden.userId !== user.id) {
+            throw new ForbiddenException('You are not allowed to update this garden');
+        }
+
+        // kiểm tra tên trùng lặp
+        const duplicate = await this.prisma.garden.findFirst({
+            where:{
+                name: dto.name,
+                userId: garden.userId
+            }
+        })
+
+        if (duplicate) {
+            throw new BadRequestException('Garden name already exists for this user');
+        }
+
         return await this.prisma.garden.update({
             where: {id},
             data: dto
@@ -71,10 +103,14 @@ export class GardenService{
     }
 
     // delete garden 
-    async deleteGarden(id: number){
+    async deleteGarden(id: number, user: any ){
         const garden = await this.prisma.garden.findUnique({where:{id}})
         if(!garden){
             throw new NotFoundException(`Not found garden: ${id}`)
+        }
+
+        if(user.role !== 'admin' && user.id !== garden.userId){
+            throw new ForbiddenException('You are not allowed to delete this garden');
         }
         return await this.prisma.garden.delete({
             where: {id},
